// Project/InternalAuditSystem.Web/ClientApp/src/app/swaggerapi/AngularFiles/model/auditPlanStatus.ts
/**
 * My Title
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export type AuditPlanStatus = 0 | 1 | 2;

export const AuditPlanStatus = {
    NUMBER_0: 0 as AuditPlanStatus,
    NUMBER_1: 1 as AuditPlanStatus,
    NUMBER_2: 2 as AuditPlanStatus
};