// Project/InternalAuditSystem.Web/ClientApp/src/app/swaggerapi/AngularFiles/model/auditPlanSectionType.ts
/**
 * My Title
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export type AuditPlanSectionType = 0 | 1 | 2 | 3 | 4;

export const AuditPlanSectionType = {
    NUMBER_0: 0 as AuditPlanSectionType,
    NUMBER_1: 1 as AuditPlanSectionType,
    NUMBER_2: 2 as AuditPlanSectionType,
    NUMBER_3: 3 as AuditPlanSectionType,
    NUMBER_4: 4 as AuditPlanSectionType
};